apply rest of @fedekunze review suggestions
