+++ x/ibc-transfer/types/keys.go
@@ -1,8 +1,6 @@
 package types
 
 import (
-	"fmt"
-
 	"github.com/tendermint/tendermint/crypto"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -39,25 +37,4 @@ var PortKey = []byte{0x01}
 // unique.
 func GetEscrowAddress(portID, channelID string) sdk.AccAddress {
 	return sdk.AccAddress(crypto.AddressHash([]byte(portID + channelID)))
-}
-
-// GetDenomPrefix returns the receiving denomination prefix
-func GetDenomPrefix(portID, channelID string) string {
-	return fmt.Sprintf("%s/%s/", portID, channelID)
-}
-
-// GetPrefixedDenom returns the denomination with the portID and channelID prefixed
-func GetPrefixedDenom(portID, channelID, baseDenom string) string {
-	return fmt.Sprintf("%s/%s/%s", portID, channelID, baseDenom)
-}
-
-// GetPrefixedCoin creates a copy of the given coin with the prefixed denom
-func GetPrefixedCoin(portID, channelID string, coin sdk.Coin) sdk.Coin {
-	return sdk.NewCoin(GetPrefixedDenom(portID, channelID, coin.Denom), coin.Amount)
-}
-
-// GetTransferCoin creates a transfer coin with the port ID and channel ID
-// prefixed to the base denom.
-func GetTransferCoin(portID, channelID, baseDenom string, amount int64) sdk.Coin {
-	return sdk.NewInt64Coin(GetPrefixedDenom(portID, channelID, baseDenom), amount)
 }
