+++ x/ibc-transfer/types/msgs.go
@@ -63,10 +63,11 @@ func (msg MsgTransfer) ValidateBasic() error {
 	}
 
 	// sanity check that validate basic on fungible token packet passes
-	// NOTE: this should always pass since validation checks should be the same
+	// NOTE: this should always pass since validation checks should be the
+	// same. Please open an issue if you encounter an error on this line.
 	packet := NewFungibleTokenPacketData(msg.Token, msg.Sender.String(), msg.Receiver, msg.Source)
 	if err := packet.ValidateBasic(); err != nil {
-		return sdkerrors.Wrap(err, "unexpected failure in fungible token packet validation, please open an issue")
+		return err
 	}
 	return nil
 }
